ice breaker/requirements.txt
python-dotenv
langchainhub
langchain
google-search-results
